move tag interface to notes index file
