--- src/app/domain/notes/components/create-note/components/add-tags/index.ts
@@ -1,5 +1,0 @@
-export interface Tag {
-  name: string;
-  id: string;
-  updatedAt: string;
-}
